function-app/requirements.txt
azure-functions==1.18.0
azure-cosmos==4.5.1
pytest